requirements.txt
ultralytics==8.3.62  
torch>=1.8.0  
torchvision>=0.9.0  
opencv-python  
numpy  
Pillow  
streamlit  
